Push

+++ Assets/Scripts/NeuralNetworks/Edge.cs
@@ -6,7 +6,7 @@ using UnityEngine;
 [Serializable]
 public class Edge
 {
-    public float Weight = 0;
+    public double Weight = 0;
     public Edge()
     {
         Weight = UnityEngine.Random.Range(-1f, 1f);
